requirements.txt
streamlit==1.40.0
pandas
azure-storage-blob
python-dotenv
plotly